Move the `CombinedLoader` to an utility file
